requirements.txt
wheel
python-dotenv
documentcloud
nltk
